Move isValidIP helper to an utils module
