+++ src/components/IPInput.js
@@ -20,38 +20,12 @@
  */
 
 import React, { useState } from 'react';
-import { isValid, parseCIDR } from "ipaddr.js";
+import { isValidIP } from '../lib/utils';
 import { TextInput, ValidatedOptions } from '@patternfly/react-core';
 
 import cockpit from 'cockpit';
 
 const _ = cockpit.gettext;
-
-/**
- *
- * Check if an IP is valid
- *
- * Sadly, the ipaddr.js library do not validate IPs with "192.168.1.1/32"
- * notation.
- *
- *    * https://github.com/whitequark/ipaddr.js/issues/13
- *    * https://github.com/whitequark/ipaddr.js/issues/25
- *
- * @param {string} value - An IP Address
- * @return {boolean} true if given IP is valid; false otherwise.
- */
-const isValidIP = (value) => {
-    const [ip, cidr] = value.split("/");
-
-    if (!cidr) return isValid(ip);
-
-    try {
-        parseCIDR(value);
-        return true;
-    } catch {
-        return false;
-    }
-};
 
 const IPInput = ({ placeholder, onError, ...props }) => {
     const [validated, setValidated] = useState("default");
